[Movie] Move "MediaStatusColumn" to MovieModel
The columns are only used for the movie model at the moment.
